+++ HISTORY.md
@@ -4,6 +4,7 @@ History of features added to LiChess Tools in time (reversed order)
 
 - previous study menu items count
 - highest rating/sandbagging warning
+- opponents page scores
 
 17 Mar